Refactor ChatList and ChatInput
