+++ src/pages/Login.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 import firebase from "firebase/app";
-import db from "../config/dbFirebase";
 import StyledFirebaseAuth from "react-firebaseui/StyledFirebaseAuth";
 
 const Login = () => {
