+++ src/pages/PlayRoom.jsx
@@ -1,10 +1,10 @@
-import React, { useState, useEffect, useRef } from "react";
+import React, { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
 import { makeStyles } from "@material-ui/core/styles";
-import { Grid, Paper, TextField, Button, Typography } from "@material-ui/core";
-import SendIcon from "@material-ui/icons/Send";
+import { Grid, Paper, Typography } from "@material-ui/core";
 import Container from "../components/Container";
-import ChatItem from "../components/ChatItem";
+import ChatList from "../components/ChatList";
+import ChatInput from "../components/ChatInput";
 import firebase from "firebase/app";
 import db from "../config/dbFirebase";
 
@@ -12,28 +12,16 @@ const styles = makeStyles({
   root: {
     height: "90%",
   },
-  chatContainer: {
-    height: "60vh",
-    overflow: "auto",
-  },
-  chat: {
-    height: "100%",
-    overflow: "auto",
-    background: "#DDD",
-  },
 });
 const background =
   "radial-gradient(circle, rgba(2,0,36,1) 0%, rgba(9,9,121,1) 35%, rgba(0,212,255,1) 100%)";
 
 export function PlayRoom(props) {
   const [room, setRoom] = useState({});
-  const [message, setMessage] = useState("");
   const [messages, setMessages] = useState([]);
 
   const { id } = useParams();
   const classes = styles(props);
-
-  const refChatPaper = useRef();
 
   useEffect(() => {
     db.collection("rooms")
@@ -58,12 +46,10 @@ export function PlayRoom(props) {
           });
         });
         setMessages(auxMessages);
-        refChatPaper.current.scrollTop = refChatPaper.current.scrollHeight;
       });
-  }, []);
+  }, [id]);
 
-  const handleSubmitMessage = (e) => {
-    e.preventDefault();
+  const handleSubmitMessage = (message) => {
     db.collection("rooms")
       .doc(id)
       .collection("messages")
@@ -75,7 +61,6 @@ export function PlayRoom(props) {
         },
         created: firebase.firestore.FieldValue.serverTimestamp(),
       });
-    setMessage("");
   };
 
   return (
@@ -93,64 +78,8 @@ export function PlayRoom(props) {
             </Paper>
           </Grid>
         </Grid>
-
-        <Grid item xs={12} className={classes.chatContainer} spacing={2}>
-          <Paper elevation={2} className={classes.chat} ref={refChatPaper}>
-            <Grid container xs={12} direction="column" spacing={1}>
-              {messages.map((msg) => (
-                <ChatItem msg={msg} />
-              ))}
-            </Grid>
-          </Paper>
-        </Grid>
-
-        <Grid item xs={12}>
-          <Grid container direction="column" spacing={1}>
-            <Grid item>
-              <Paper elevation={2}>Botonera</Paper>
-            </Grid>
-            <Grid item>
-              <form onSubmit={handleSubmitMessage}>
-                <Paper elevation={2}>
-                  <Grid
-                    item
-                    container
-                    xs={12}
-                    direction="row"
-                    justify="space-between"
-                    spacing={1}
-                    alignItems="center"
-                  >
-                    <Grid item xs={10}>
-                      <TextField
-                        id="chatInput"
-                        label="Mensaje"
-                        variant="outlined"
-                        size="small"
-                        fullWidth
-                        onChange={(e) => setMessage(e.target.value)}
-                        value={message}
-                      />
-                    </Grid>
-                    <Grid item xs={2}>
-                      <Button
-                        variant="contained"
-                        color="primary"
-                        className={classes.button}
-                        endIcon={<SendIcon />}
-                        size="large"
-                        fullWidth
-                        type="submit"
-                      >
-                        Enviar
-                      </Button>
-                    </Grid>
-                  </Grid>
-                </Paper>
-              </form>
-            </Grid>
-          </Grid>
-        </Grid>
+        <ChatList messages={messages} />
+        <ChatInput onSubmitMessage={handleSubmitMessage} />
       </Grid>
     </Container>
   );
